[rdrawable] hide many details in source files
Only items for dictionary generation will remain in headers.
Idea that Browsable::RProvider and DrawProvider should all necessary
API.
