+++ gui/browserv7/src/RBrowsableTDirectory.cxx
@@ -8,3 +8,265 @@
 
 
 #include <ROOT/RBrowsableTDirectory.hxx>
+
+#include "ROOT/RLogger.hxx"
+
+#include "TSystem.h"
+#include "TKey.h"
+#include "TDirectory.h"
+#include "TROOT.h"
+#include "TFile.h"
+#include "TH1.h"
+
+using namespace std::string_literals;
+
+using namespace ROOT::Experimental;
+using namespace ROOT::Experimental::Browsable;
+
+
+// ===============================================================================================================
+
+
+class TDirectoryLevelIter : public RLevelIter {
+   TDirectory *fDir{nullptr};         ///<! current directory handle
+   std::unique_ptr<TIterator>  fIter; ///<! created iterator
+   TKey *fKey{nullptr};               ///<! currently selected key
+   std::string fCurrentName;          ///<! current key name
+
+   bool CreateIter()
+   {
+      if (!fDir) return false;
+      fIter.reset(fDir->GetListOfKeys()->MakeIterator());
+      fKey = nullptr;
+      return true;
+   }
+
+   void CloseIter()
+   {
+      fIter.reset(nullptr);
+      fKey = nullptr;
+   }
+
+   bool NextDirEntry()
+   {
+      fCurrentName.clear();
+      if (!fIter) return false;
+
+      fKey = dynamic_cast<TKey *>(fIter->Next());
+
+      if (!fKey) {
+         fIter.reset();
+         return false;
+      }
+
+      fCurrentName = fKey->GetName();
+      fCurrentName.append(";");
+      fCurrentName.append(std::to_string(fKey->GetCycle()));
+
+      return true;
+   }
+
+public:
+   explicit TDirectoryLevelIter(TDirectory *dir) : fDir(dir) { CreateIter(); }
+
+   virtual ~TDirectoryLevelIter() = default;
+
+   bool Reset() override { return CreateIter(); }
+
+   bool Next() override { return NextDirEntry(); }
+
+   // use default implementation for now
+   // bool Find(const std::string &name) override { return FindDirEntry(name); }
+
+   bool HasItem() const override { return fKey && !fCurrentName.empty(); }
+
+   std::string GetName() const override { return fCurrentName; }
+
+   int CanHaveChilds() const override
+   {
+      std::string clname = fKey->GetClassName();
+      return (clname.find("TDirectory") == 0) ? 1 : 0;
+   }
+
+   std::string GetClassIcon(const std::string &classname)
+   {
+      if (classname == "TTree" || classname == "TNtuple")
+         return "sap-icon://tree"s;
+      if (classname == "TDirectory" || classname == "TDirectoryFile")
+         return "sap-icon://folder-blank"s;
+
+      return "sap-icon://electronic-medical-record"s;
+   }
+
+   /** Create element for the browser */
+   std::unique_ptr<RBrowserItem> CreateBrowserItem() override
+   {
+      auto item = std::make_unique<RBrowserTKeyItem>(GetName(), CanHaveChilds());
+
+      item->className = fKey->GetClassName();
+
+      item->SetIcon(GetClassIcon(item->className));
+
+      return item;
+   }
+
+   /** Returns full information for current element */
+   std::shared_ptr<RElement> GetElement() override;
+
+};
+
+// ===============================================================================================================
+
+
+
+class TKeyElement : public RElement {
+   TDirectory *fDir{nullptr};
+   TKey *fKey{nullptr};
+
+public:
+   TKeyElement(TDirectory *dir, TKey *key) : fDir(dir), fKey(key) {}
+
+   virtual ~TKeyElement() = default;
+
+   /** Name of RBrowsable, must be provided in derived classes */
+   std::string GetName() const override
+   {
+      std::string name = fKey->GetName();
+      name.append(";");
+      name.append(std::to_string(fKey->GetCycle()));
+      return name;
+   }
+
+   /** Title of RBrowsable (optional) */
+   std::string GetTitle() const override { return fKey->GetTitle(); }
+
+   /** Create iterator for childs elements if any */
+   std::unique_ptr<RLevelIter> GetChildsIter() override
+   {
+      std::string clname = fKey->GetClassName();
+
+      if (clname.find("TDirectory") != 0)
+         return nullptr;
+
+      auto subdir = fDir->GetDirectory(GetName().c_str());
+      if (subdir)
+         return std::make_unique<TDirectoryLevelIter>(subdir);
+
+      return nullptr;
+   }
+
+   /** Return TObject depending from kind of requested result */
+   std::unique_ptr<RObject> GetObject(bool plain = false) override
+   {
+      std::string clname = fKey->GetClassName();
+
+      TObject *obj = fKey->ReadObj();
+      if (!obj) return nullptr;
+
+      if (plain)
+         return std::make_unique<RTObjectHolder>(obj);
+
+      if (obj->InheritsFrom(TH1::Class()))
+         static_cast<TH1 *>(obj)->SetDirectory(nullptr);
+
+      if (fDir) fDir->Remove(obj); // remove object while ownreship will be delivered to return value
+
+      return std::make_unique<RUnique<TObject>>(obj);
+   }
+
+};
+
+// ==============================================================================================
+
+
+std::shared_ptr<RElement> TDirectoryLevelIter::GetElement()
+{
+   return std::make_shared<TKeyElement>(fDir, fKey);
+}
+
+// ==============================================================================================
+
+class TDirectoryElement : public RElement {
+   std::string fFileName;       ///<!   file name
+   TDirectory *fDir{nullptr};   ///<!   subdirectory (ifany)
+
+   ///////////////////////////////////////////////////////////////////
+   /// Get TDirectory. Checks if parent file is still there. If not, means it was closed outside ROOT
+
+   TDirectory *GetDir()
+   {
+      if (fDir) {
+         if (!gROOT->GetListOfFiles()->FindObject(fDir->GetFile()))
+            fDir = nullptr;
+      } else if (!fFileName.empty()) {
+         fDir = TFile::Open(fFileName.c_str());
+      }
+
+      return fDir;
+   }
+
+public:
+
+   TDirectoryElement(const std::string &fname, TDirectory *dir = nullptr)
+   {
+      fFileName = fname;
+      fDir = dir;
+   }
+
+   virtual ~TDirectoryElement() = default;
+
+   /** Name of RBrowsable, must be provided in derived classes */
+   std::string GetName() const override
+   {
+      if (fDir)
+         return fDir->GetName();
+
+      if (!fFileName.empty()) {
+         auto pos = fFileName.rfind("/");
+         return ((pos == std::string::npos) || (pos > fFileName.length() - 2)) ? fFileName : fFileName.substr(pos + 1);
+      }
+
+      return ""s;
+   }
+
+   /** Title of RBrowsable (optional) */
+   std::string GetTitle() const override
+   {
+      if (fDir)
+         return fDir->GetTitle();
+
+      return "ROOT file "s + fFileName;
+   }
+
+   std::unique_ptr<RLevelIter> GetChildsIter() override
+   {
+      auto dir = GetDir();
+
+      return dir ? std::make_unique<TDirectoryLevelIter>(dir) : nullptr;
+   }
+};
+
+
+// ==============================================================================================
+
+
+class RTFileProvider : public RProvider {
+protected:
+
+   std::shared_ptr<RElement> DoOpenFile(const std::string &fullname) const override
+   {
+      auto f = TFile::Open(fullname.c_str());
+      if (!f) return nullptr;
+
+      return std::make_shared<TDirectoryElement>(fullname, f);
+   }
+
+};
+
+struct RTFileProviderReg {
+   std::shared_ptr<RTFileProvider> provider;
+   RTFileProviderReg() { provider = std::make_shared<RTFileProvider>(); RProvider::RegisterFile("root", provider); }
+   ~RTFileProviderReg() { RProvider::Unregister(provider); }
+} newRTFileProviderReg;
+
+
